test(nextcloud-talk): align CI fixtures
